Implements sys package for the following ticket:  - MDEV-34331: System package SYS.UTL_I18N
The package is a set of services that provides additional globalization
functionality.

The ANY_CS character set was also implemented so that the character set set
by "set names" command or the character set of the database as the character
set of the variable set with ANY_CS, especially for the transliterate
function.

Added the following functions:
- TRANSLITERATE - Perform script transliteration
- RAW_TO_CHAR - Convert RAW data from a valid character set to a VARCHAR2
string in the database character set
- STRING_TO_RAW - Convert a VARCHAR2 string to another valid character set
and returns the result as RAW data